sw_redlinehide_2: view cursor movement, SwCursorShell::LeftRight()
Just put a loop in SwCursor::LeftRight() to repeat a movement if
it didn't actually advance the text frame index; the
SwContentNode::GoPrevious()/GoNext() take into account hidden text
attributes from SwScriptInfo, so this ought to result in end
positions that are neither hidden by attributes nor by redlines.

This requires passing the layout to SwCursor.

Change-Id: Ieb623840f6390fa6f1c78b7458ad8dc6523a2744